pyproject.toml
[project]
name = "teleme-chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.20.14",
    "openai>=1.57.3",
    "pydantic-settings>=2.7.0",
    "python-fasthtml>=0.10.1",
    "qdrant-client>=1.12.1",
    "ruff>=0.8.2",
    "sentence-transformers>=3.3.1",
]
